Moved internal/otlpjson to loggingexporter/internal/jsonstream
