+++ exporter/loggingexporter/internal/jsonstream/metrics_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package otlpjson
+package jsonstream
 
 import (
 	"testing"
